# back-front/src/main/java/com/dev/strangerred/dto/CommentsDto.java
package com.dev.strangerred.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import java.time.Instant;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class CommentsDto {

    private Long id;
    private Long postId;
    private Instant createdDate;
    private String text;
    private String userName;
}