# back-front/src/main/java/com/dev/strangerred/dto/LoginRequest.java
package com.dev.strangerred.dto;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class LoginRequest {

    private String username;
    private String password;
}