# back-front/src/main/java/com/dev/strangerred/dto/VoteDto.java
package com.dev.strangerred.dto;

import com.dev.strangerred.model.VoteType;
import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class VoteDto {

    private VoteType voteType;
    private Long postId;
}