# back-front/src/main/resources/SwaggerStrangerRed.json
{
	"swagger": "2.0",
	"info": {
		"description": "API for Stranger Red Social Network Application",
		"version": "1.0",
		"title": "Stranger Red Social Network API",
		"contact": {
			"name": "Hemerson Carlos Cabral",
			"email": "hemersoncarlos93@hotmail.com"
		},
		"license": {
			"name": "Apache License Version 2.0"
		}
	},
	"host": "localhost:8080",
	"basePath": "/",
	"tags": [
		{
			"name": "auth-controller",
			"description": "Auth Controller"
		},
		{
			"name": "basic-error-controller",
			"description": "Basic Error Controller"
		},
		{
			"name": "comments-controller",
			"description": "Comments Controller"
		},
		{
			"name": "post-controller",
			"description": "Post Controller"
		},
		{
			"name": "subreddit-controller",
			"description": "Subreddit Controller"
		},
		{
			"name": "vote-controller",
			"description": "Vote Controller"
		}
	],
	"paths": {
		"/api/auth/accountVerification/{token}": {
			"get": {
				"tags": [
					"auth-controller"
				],
				"summary": "verifyAccount",
				"operationId": "verifyAccountUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "token",
						"in": "path",
						"description": "token",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/auth/login": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"summary": "login",
				"operationId": "loginUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "loginRequest",
						"description": "loginRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/LoginRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/AuthenticationResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/auth/logout": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"summary": "logout",
				"operationId": "logoutUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "refreshTokenRequest",
						"description": "refreshTokenRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RefreshTokenRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/auth/refresh/token": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"summary": "refreshTokens",
				"operationId": "refreshTokensUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "refreshTokenRequest",
						"description": "refreshTokenRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RefreshTokenRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/AuthenticationResponse"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/auth/signup": {
			"post": {
				"tags": [
					"auth-controller"
				],
				"summary": "signup",
				"operationId": "signupUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "registerRequest",
						"description": "registerRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/RegisterRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "string"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/comments/": {
			"post": {
				"tags": [
					"comments-controller"
				],
				"summary": "createComment",
				"operationId": "createCommentUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "commentsDto",
						"description": "commentsDto",
						"required": true,
						"schema": {
							"$ref": "#/definitions/CommentsDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/comments/by-post/{postId}": {
			"get": {
				"tags": [
					"comments-controller"
				],
				"summary": "getAllCommentsForPost",
				"operationId": "getAllCommentsForPostUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "postId",
						"in": "path",
						"description": "postId",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/CommentsDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/comments/by-user/{userName}": {
			"get": {
				"tags": [
					"comments-controller"
				],
				"summary": "getAllCommentsForUser",
				"operationId": "getAllCommentsForUserUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "userName",
						"in": "path",
						"description": "userName",
						"required": true,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/CommentsDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/posts/": {
			"get": {
				"tags": [
					"post-controller"
				],
				"summary": "getAllPosts",
				"operationId": "getAllPostsUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PostResponse"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"post": {
				"tags": [
					"post-controller"
				],
				"summary": "createPost",
				"operationId": "createPostUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "postRequest",
						"description": "postRequest",
						"required": true,
						"schema": {
							"$ref": "#/definitions/PostRequest"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/posts/by-subreddit/{id}": {
			"get": {
				"tags": [
					"post-controller"
				],
				"summary": "getPostsBySubreddit",
				"operationId": "getPostsBySubredditUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"description": "id",
						"required": false,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PostResponse"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/posts/by-user/{name}": {
			"get": {
				"tags": [
					"post-controller"
				],
				"summary": "getPostsByUsername",
				"operationId": "getPostsByUsernameUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "username",
						"in": "query",
						"description": "username",
						"required": false,
						"type": "string"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/PostResponse"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/posts/{id}": {
			"get": {
				"tags": [
					"post-controller"
				],
				"summary": "getPost",
				"operationId": "getPostUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/PostResponse"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/subreddit": {
			"get": {
				"tags": [
					"subreddit-controller"
				],
				"summary": "getAllSubreddits",
				"operationId": "getAllSubredditsUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "array",
							"items": {
								"$ref": "#/definitions/SubredditDto"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"post": {
				"tags": [
					"subreddit-controller"
				],
				"summary": "createSubreddit",
				"operationId": "createSubredditUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "subredditDto",
						"description": "subredditDto",
						"required": true,
						"schema": {
							"$ref": "#/definitions/SubredditDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/SubredditDto"
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/subreddit/{id}": {
			"get": {
				"tags": [
					"subreddit-controller"
				],
				"summary": "getSubreddit",
				"operationId": "getSubredditUsingGET",
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"name": "id",
						"in": "path",
						"description": "id",
						"required": true,
						"type": "integer",
						"format": "int64"
					}
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"$ref": "#/definitions/SubredditDto"
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/api/votes/": {
			"post": {
				"tags": [
					"vote-controller"
				],
				"summary": "vote",
				"operationId": "voteUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"parameters": [
					{
						"in": "body",
						"name": "voteDto",
						"description": "voteDto",
						"required": true,
						"schema": {
							"$ref": "#/definitions/VoteDto"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			}
		},
		"/error": {
			"get": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingGET",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"head": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingHEAD",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			},
			"post": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingPOST",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"put": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingPUT",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"201": {
						"description": "Created"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					},
					"404": {
						"description": "Not Found"
					}
				}
			},
			"delete": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingDELETE",
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			},
			"options": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingOPTIONS",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			},
			"patch": {
				"tags": [
					"basic-error-controller"
				],
				"summary": "error",
				"operationId": "errorUsingPATCH",
				"consumes": [
					"application/json"
				],
				"produces": [
					"*/*"
				],
				"responses": {
					"200": {
						"description": "OK",
						"schema": {
							"type": "object",
							"additionalProperties": {
								"type": "object"
							}
						}
					},
					"204": {
						"description": "No Content"
					},
					"401": {
						"description": "Unauthorized"
					},
					"403": {
						"description": "Forbidden"
					}
				}
			}
		}
	},
	"definitions": {
		"AuthenticationResponse": {
			"type": "object",
			"properties": {
				"authenticationToken": {
					"type": "string"
				},
				"expiresAt": {
					"type": "string",
					"format": "date-time"
				},
				"refreshToken": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			},
			"title": "AuthenticationResponse"
		},
		"CommentsDto": {
			"type": "object",
			"properties": {
				"createdDate": {
					"type": "string",
					"format": "date-time"
				},
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"postId": {
					"type": "integer",
					"format": "int64"
				},
				"text": {
					"type": "string"
				},
				"userName": {
					"type": "string"
				}
			},
			"title": "CommentsDto"
		},
		"LoginRequest": {
			"type": "object",
			"properties": {
				"password": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			},
			"title": "LoginRequest"
		},
		"ModelAndView": {
			"type": "object",
			"properties": {
				"empty": {
					"type": "boolean"
				},
				"model": {
					"type": "object"
				},
				"modelMap": {
					"type": "object",
					"additionalProperties": {
						"type": "object"
					}
				},
				"reference": {
					"type": "boolean"
				},
				"status": {
					"type": "string",
					"enum": [
						"ACCEPTED",
						"ALREADY_REPORTED",
						"BAD_GATEWAY",
						"BAD_REQUEST",
						"BANDWIDTH_LIMIT_EXCEEDED",
						"CHECKPOINT",
						"CONFLICT",
						"CONTINUE",
						"CREATED",
						"DESTINATION_LOCKED",
						"EXPECTATION_FAILED",
						"FAILED_DEPENDENCY",
						"FORBIDDEN",
						"FOUND",
						"GATEWAY_TIMEOUT",
						"GONE",
						"HTTP_VERSION_NOT_SUPPORTED",
						"IM_USED",
						"INSUFFICIENT_SPACE_ON_RESOURCE",
						"INSUFFICIENT_STORAGE",
						"INTERNAL_SERVER_ERROR",
						"I_AM_A_TEAPOT",
						"LENGTH_REQUIRED",
						"LOCKED",
						"LOOP_DETECTED",
						"METHOD_FAILURE",
						"METHOD_NOT_ALLOWED",
						"MOVED_PERMANENTLY",
						"MOVED_TEMPORARILY",
						"MULTIPLE_CHOICES",
						"MULTI_STATUS",
						"NETWORK_AUTHENTICATION_REQUIRED",
						"NON_AUTHORITATIVE_INFORMATION",
						"NOT_ACCEPTABLE",
						"NOT_EXTENDED",
						"NOT_FOUND",
						"NOT_IMPLEMENTED",
						"NOT_MODIFIED",
						"NO_CONTENT",
						"OK",
						"PARTIAL_CONTENT",
						"PAYLOAD_TOO_LARGE",
						"PAYMENT_REQUIRED",
						"PERMANENT_REDIRECT",
						"PRECONDITION_FAILED",
						"PRECONDITION_REQUIRED",
						"PROCESSING",
						"PROXY_AUTHENTICATION_REQUIRED",
						"REQUESTED_RANGE_NOT_SATISFIABLE",
						"REQUEST_ENTITY_TOO_LARGE",
						"REQUEST_HEADER_FIELDS_TOO_LARGE",
						"REQUEST_TIMEOUT",
						"REQUEST_URI_TOO_LONG",
						"RESET_CONTENT",
						"SEE_OTHER",
						"SERVICE_UNAVAILABLE",
						"SWITCHING_PROTOCOLS",
						"TEMPORARY_REDIRECT",
						"TOO_EARLY",
						"TOO_MANY_REQUESTS",
						"UNAUTHORIZED",
						"UNAVAILABLE_FOR_LEGAL_REASONS",
						"UNPROCESSABLE_ENTITY",
						"UNSUPPORTED_MEDIA_TYPE",
						"UPGRADE_REQUIRED",
						"URI_TOO_LONG",
						"USE_PROXY",
						"VARIANT_ALSO_NEGOTIATES"
					]
				},
				"view": {
					"$ref": "#/definitions/View"
				},
				"viewName": {
					"type": "string"
				}
			},
			"title": "ModelAndView"
		},
		"PostRequest": {
			"type": "object",
			"properties": {
				"description": {
					"type": "string"
				},
				"postId": {
					"type": "integer",
					"format": "int64"
				},
				"postName": {
					"type": "string"
				},
				"subredditName": {
					"type": "string"
				},
				"url": {
					"type": "string"
				}
			},
			"title": "PostRequest"
		},
		"PostResponse": {
			"type": "object",
			"properties": {
				"commentCount": {
					"type": "integer",
					"format": "int32"
				},
				"description": {
					"type": "string"
				},
				"downVote": {
					"type": "boolean"
				},
				"duration": {
					"type": "string"
				},
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"postName": {
					"type": "string"
				},
				"subredditName": {
					"type": "string"
				},
				"upVote": {
					"type": "boolean"
				},
				"url": {
					"type": "string"
				},
				"userName": {
					"type": "string"
				},
				"voteCount": {
					"type": "integer",
					"format": "int32"
				}
			},
			"title": "PostResponse"
		},
		"RefreshTokenRequest": {
			"type": "object",
			"required": [
				"refreshToken"
			],
			"properties": {
				"refreshToken": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			},
			"title": "RefreshTokenRequest"
		},
		"RegisterRequest": {
			"type": "object",
			"properties": {
				"email": {
					"type": "string"
				},
				"password": {
					"type": "string"
				},
				"username": {
					"type": "string"
				}
			},
			"title": "RegisterRequest"
		},
		"SubredditDto": {
			"type": "object",
			"properties": {
				"description": {
					"type": "string"
				},
				"id": {
					"type": "integer",
					"format": "int64"
				},
				"name": {
					"type": "string"
				},
				"numberOfPosts": {
					"type": "integer",
					"format": "int32"
				}
			},
			"title": "SubredditDto"
		},
		"View": {
			"type": "object",
			"properties": {
				"contentType": {
					"type": "string"
				}
			},
			"title": "View"
		},
		"VoteDto": {
			"type": "object",
			"properties": {
				"postId": {
					"type": "integer",
					"format": "int64"
				},
				"voteType": {
					"type": "string",
					"enum": [
						"DOWNVOTE",
						"UPVOTE"
					]
				}
			},
			"title": "VoteDto"
		}
	}
}